Minor improvements to script, version increment

+++ variable-populator.sh
@@ -11,17 +11,23 @@ logit() {
 # Populate platform manifests with environment variables
 varpop() {
 	cd /tmp
-	git clone https://$GITHUB_TOKEN@github.com/weavegitops/$CLUSTER_NAME.git
-	logit "WGE repo cloned"
-	cd $CLUSTER_NAME
+	git clone https://${GITHUB_TOKEN}@github.com/weavegitops/${CLUSTER_NAME}.git
+	logit "WGE repo cloned, check deployment status"
+	cd ${CLUSTER_NAME}
+	if [ -f platform/.done ]
+	then
+		logit "variables previously populated, exiting"
+		exit 0
+	fi
 	cp -R bin/platform/* platform
 	logit "Platform directory copied"
 	for file in $(ls platform/)
 	do
-		envsubst < platform/$file > /tmp/$file && mv /tmp/$file platform/$file
-		logit "Variables replaced in $file"
+		envsubst < platform/${file} > /tmp/${file} && mv /tmp/${file} platform/${file}
+		logit "Variables replaced in ${file}"
 	done
 	logit "Variables replaced in all files in platform directory"
+	touch platform/.done
 	git config user.name github-actions
 	git config user.email github-actions@github.com
 	git add platform/.
@@ -35,16 +41,16 @@ varpop() {
 fail=0
 until [ $fail -gt 9 ]
 do
-	if [ "$(kubectl get terraform --namespace $NAMESPACE $CLUSTER_NAME-wge-repo -o jsonpath='{.status.conditions[0].status}')" == "True" ]
+	if [ "$(kubectl get terraform --namespace ${NAMESPACE} ${CLUSTER_NAME}-wge-repo -o jsonpath='{.status.conditions[0].status}')" == "True" ]
 	then
 		logit "WGE repo provisioned, populating repo"
 		varpop
 		exit 0
 	else
 		logit "WGE repo not yet provisioned, waiting 2 minutes"
-		#sleep 120
-		((fail++))
-		echo $fail
+		sleep 120
+		(( ++fail ))
+		echo ${fail}
 	fi
 done
